remove dependency cycle and linter errors
